Command class is now a member class of Commandevent
